+++ src/NotifierServiceProvider.php
@@ -17,6 +17,8 @@ class NotifierServiceProvider extends ServiceProvider implements DeferrableProvi
         $this->registerMailer();
 
         $this->registerNotifier();
+
+        $this->registerIlluminateMailerResolver();
     }
 
     /**
@@ -27,10 +29,11 @@ class NotifierServiceProvider extends ServiceProvider implements DeferrableProvi
     protected function registerMailer(): void
     {
         $this->app->singleton('orchestra.mail', function ($app) {
-            $mailer = new Mailer($app, new TransportManager($app));
+            $mailer = new Mailer($app, $transport = new TransportManager($app));
 
             if ($app->bound('orchestra.platform.memory')) {
-                $mailer->attach($app->make('orchestra.platform.memory'));
+                $mailer->attach($memory = $app->make('orchestra.platform.memory'));
+                $transport->setMemoryProvider($memory);
             }
 
             if ($app->bound('queue')) {
@@ -50,6 +53,18 @@ class NotifierServiceProvider extends ServiceProvider implements DeferrableProvi
     {
         $this->app->singleton('orchestra.notifier', function ($app) {
             return new NotifierManager($app);
+        });
+    }
+
+    /**
+     * Register the service provider for notifier.
+     *
+     * @return void
+     */
+    protected function registerIlluminateMailerResolver(): void
+    {
+        $this->app->afterResolving('mailer', function ($service) {
+            $this->app->make('orchestra.mail')->configureIlluminateMailer($service);
         });
     }